Refactor snapshot tests to remove unnecessary fields
- Removed `tool_calls`, `model`, and `reasoning_details` from various snapshot test cases to streamline the data structure.
- Updated multiple snapshot files in `forge_domain` and `forge_provider` to reflect these changes, ensuring consistency across tests.
- Adjusted assertions in tests to focus on essential message content without extraneous fields.
